+++ HISTORY.rst
@@ -7,6 +7,8 @@ develop
 - Copy processing algorithms from Models and Simulations plugin to Rana desktop client
 - Add remove from project file action that shadows delete
 - Show relative timestamps in files browser, file view and revisions view
+- Group folders on top in files browser
+- Sort folder and file names case insensitive
 
 
 1.1.21 (unreleased)